pyproject.toml
[project]
name = "trialexp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.14.0",
    "annotated-types==0.7.0",
    "anyio==4.4.0",
    "appdirs==1.4.4",
    "argon2-cffi==23.1.0",
    "argon2-cffi-bindings==21.2.0",
    "arrow==1.3.0",
    "asciitree==0.3.3",
    "asgiref==3.8.1",
    "asttokens==2.4.1",
    "async-lru==2.0.4",
    "async-timeout==5.0.1",
    "attrs==23.2.0",
    "babel==2.15.0",
    "backports-tarfile==1.2.0",
    "beautifulsoup4==4.12.3",
    "bidict==0.23.1",
    "bleach==6.1.0",
    "bokeh==3.4.3",
    "bottleneck==1.4.2",
    "build==1.2.2.post1",
    "certifi==2024.6.2",
    "cffi==1.16.0",
    "charset-normalizer==3.3.2",
    "click==8.1.7",
    "cloudpickle==3.0.0",
    "colorama==0.4.6",
    "colorcet==3.1.0",
    "comm==0.2.2",
    "configargparse>=1.7",
    "connection-pool==0.0.3",
    "contourpy==1.2.1",
    "cryptography==44.0.0",
    "cycler==0.12.1",
    "dask==2024.8.0",
    "dask-expr==1.1.10",
    "datashader==0.16.3",
    "debugpy==1.8.1",
    "decorator==5.1.1",
    "defusedxml==0.7.1",
    "distro==1.9.0",
    "docutils==0.21.2",
    "dpath==2.2.0",
    "dtaidistance==2.3.13",
    "elephant==1.1.0",
    "exceptiongroup==1.2.1",
    "executing==2.0.1",
    "faiss-cpu==1.8.0",
    "fastapi[standard]==0.115.5",
    "fastdtw==0.3.4",
    "fasteners==0.19",
    "fastjsonschema==2.20.0",
    "filelock==3.15.3",
    "fonttools==4.53.0",
    "fqdn==1.5.1",
    "fsspec==2024.6.0",
    "gitdb==4.0.11",
    "gitpython==3.1.43",
    "greenlet==3.1.1",
    "gunicorn==23.0.0",
    "h11==0.14.0",
    "h5netcdf==1.3.0",
    "h5py==3.11.0",
    "holoviews==1.19.1",
    "htmltools==0.5.2",
    "httpcore==1.0.5",
    "httpx==0.27.0",
    "humanfriendly==10.0",
    "idna==3.7",
    "importlib-metadata==7.2.0",
    "importlib-resources==6.4.0",
    "ipykernel>=6.29.4",
    "ipython==9.7.0",
    "ipywidgets==8.1.3",
    "isoduration==20.11.0",
    "jaraco-classes==3.4.0",
    "jaraco-context==6.0.1",
    "jaraco-functools==4.1.0",
    "jedi==0.19.1",
    "jeepney==0.8.0",
    "jinja2==3.1.4",
    "joblib==1.4.2",
    "json5==0.9.25",
    "jsonpointer==3.0.0",
    "jsonschema==4.22.0",
    "jsonschema-specifications==2023.12.1",
    "jupyter-client==8.6.2",
    "jupyter-core==5.7.2",
    "jupyter-events==0.10.0",
    "jupyter-lsp==2.2.5",
    "jupyter-server==2.14.1",
    "jupyter-server-mathjax==0.2.6",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.2.2",
    "jupyterlab-git==0.50.1",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.27.2",
    "jupyterlab-widgets==3.0.11",
    "keyring==25.5.0",
    "kiwisolver==1.4.5",
    "lazy-loader==0.4",
    "linkify-it-py==2.0.3",
    "llvmlite==0.43.0",
    "locket==1.0.0",
    "loguru>=0.7.2",
    "mako==1.3.6",
    "markdown==3.6",
    "markdown-it-py==3.0.0",
    "markupsafe==2.1.5",
    "matplotlib==3.9.0",
    "matplotlib-inline==0.1.7",
    "mdit-py-plugins==0.4.1",
    "mdurl==0.1.2",
    "mistune==3.0.2",
    "more-itertools==10.5.0",
    "mpmath==1.3.0",
    "multipledispatch==1.0.0",
    "nbclient==0.10.0",
    "nbconvert==7.16.4",
    "nbdime==4.0.1",
    "nbformat==5.10.4",
    "neo==0.13.1",
    "nest-asyncio==1.6.0",
    "networkx==3.2.1",
    "nh3==0.2.19",
    "notebook-shim==0.2.4",
    "numba==0.60.0",
    "numcodecs==0.12.1",
    "numpy==1.26.0",
    "open-ephys-python-tools==0.1.10",
    "overrides==7.7.0",
    "packaging==24.1",
    "pandas==2.1.4",
    "pandas-flavor==0.6.0",
    "pandocfilters==1.5.1",
    "panel==1.4.5",
    "param==2.1.1",
    "parso==0.8.4",
    "partd==1.4.2",
    "patsy==0.5.6",
    "pexpect==4.9.0",
    "pillow==10.3.0",
    "pingouin==0.5.4",
    "pipdeptree==2.16.2",
    "pkginfo==1.10.0",
    "plac==1.4.3",
    "platformdirs==4.2.2",
    "plotly==5.22.0",
    "probeinterface==0.2.23",
    "prometheus-client==0.20.0",
    "prompt-toolkit>=3.0.36",
    "psutil==6.0.0",
    "ptyprocess==0.7.0",
    "pulp==2.7.0",
    "pure-eval==0.2.2",
    "pyarrow==17.0.0",
    "pycparser==2.22",
    "pyct==0.5.0",
    "pydantic==2.10.2",
    "pydantic-core==2.27.1",
    "pygments>=2.18.0",
    "pymer4==0.9.2",
    "pyparsing==3.1.2",
    "pyproject-hooks==1.2.0",
    "pyqt6==6.7.0",
    "pyqt6-qt6==6.7.1",
    "pyqt6-sip==13.6.0",
    "pyqtgraph==0.13.7",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-engineio==4.10.1",
    "python-json-logger==2.0.7",
    "python-multipart==0.0.9",
    "python-socketio==5.11.4",
    "pytz==2024.1",
    "pyviz-comms==3.0.3",
    "pyyaml==6.0.1",
    "pyzmq==26.0.3",
    "qtpy==2.4.1",
    "quantities==0.15.0",
    "questionary>=2.0.1",
    "readme-renderer==44.0",
    "redis==5.2.0",
    "referencing==0.35.1",
    "reflex==0.6.6",
    "reflex-chakra==0.6.2",
    "reflex-hosting-cli==0.1.21",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "reretry==0.11.8",
    "rfc3339-validator==0.1.4",
    "rfc3986==2.0.0",
    "rfc3986-validator==0.1.1",
    "rich==13.9.4",
    "rpds-py==0.18.1",
    "rust-just==1.40.0",
    "scikit-learn==1.5.0",
    "scipy==1.13.1",
    "seaborn==0.13.2",
    "secretstorage==3.3.3",
    "send2trash==1.8.3",
    "shellingham==1.5.4",
    "shiny==0.10.2",
    "simple-websocket==1.1.0",
    "six==1.16.0",
    "smart-open==7.0.4",
    "smmap==5.0.1",
    "sniffio==1.3.1",
    "sonpy==1.9.5",
    "soupsieve==2.5",
    "spikeinterface==0.101.0",
    "sqlalchemy==2.0.36",
    "sqlmodel==0.0.22",
    "stack-data==0.6.3",
    "starlette==0.41.3",
    "starlette-admin==0.14.1",
    "statsmodels==0.14.2",
    "stopit==1.1.2",
    "sympy==1.13.1",
    "tabulate==0.9.0",
    "tenacity==8.4.1",
    "terminado==0.18.1",
    "threadpoolctl==3.5.0",
    "throttler==1.2.2",
    "tinycss2==1.3.0",
    "tomli==2.0.1",
    "tomlkit==0.13.2",
    "toolz==0.12.1",
    "toposort==1.10",
    "tornado==6.4.1",
    "tqdm==4.66.4",
    "traitlets==5.14.3",
    "tslearn==0.6.3",
    "twine==5.1.1",
    "typer==0.14.0",
    "types-python-dateutil==2.9.0.20240316",
    "typing-extensions==4.12.2",
    "tzdata==2024.1",
    "tzlocal==5.2",
    "uc-micro-py==1.0.3",
    "uri-template==1.3.0",
    "urllib3==2.2.2",
    "uvicorn==0.30.1",
    "watchfiles==0.22.0",
    "wcwidth==0.2.13",
    "webcolors==24.6.0",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "websockets==12.0",
    "widgetsnbextension==4.0.11",
    "wrapt==1.16.0",
    "wsproto==1.2.0",
    "xarray==2024.6.0",
    "xyzservices==2024.6.0",
    "zarr==2.17.2",
    "zipp==3.19.2",
    "zmq==0.0.0",
]
[project.optional-dependencies]
full = [
    "torch",
    "torchaudio",
    "kilosort",
    "snakemake>=9.0.0",
    "snakehelper",
]

[tool.uv.sources]
snakehelper = { git = "https://github.com/teristam/snakehelper.git"}
torch = [
  { index = "pytorch-cu118", extra="full" },
]
torchaudio = [
  { index = "pytorch-cu118", extra="full"},
]


[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
